+++ Sources/Utils.swift
@@ -17,32 +17,32 @@
 import Dispatch
 import Foundation
 
-public class Collection<T> {
+class Collection<T> {
 
   internal let semaphoreQueue = DispatchSemaphore(value: 1)
   internal var list: [T]
-  public let size: Int
+  let size: Int
 
-  public var isEmpty: Bool {
+  var isEmpty: Bool {
     semaphoreQueue.wait()
     let empty = list.isEmpty
     semaphoreQueue.signal()
     return empty
   }
 
-  public var count: Int {
+  var count: Int {
     semaphoreQueue.wait()
     let count = list.count
     semaphoreQueue.signal()
     return count
   }
 
-  public init(size: Int) {
+  init(size: Int) {
     self.size = size
     self.list = [T]()
   }
 
-  public func add(_ element: T) {
+  func add(_ element: T) {
     semaphoreQueue.wait()
     list.append(element)
     if list.count > size {
@@ -51,44 +51,43 @@ public class Collection<T> {
     semaphoreQueue.signal()
   }
 
-  public func removeFirst() -> T? {
+  func removeFirst() -> T? {
     semaphoreQueue.wait()
     let element: T? = list.removeFirst()
     semaphoreQueue.signal()
     return element
   }
 
-  public func removeLast() -> T? {
+  func removeLast() -> T? {
     semaphoreQueue.wait()
     let element: T? = list.removeLast()
     semaphoreQueue.signal()
     return element
   }
 
-  public func peekFirst() -> T? {
+  func peekFirst() -> T? {
     semaphoreQueue.wait()
     let element: T? = list.first
     semaphoreQueue.signal()
     return element
   }
 
-  public func peekLast() -> T? {
+  func peekLast() -> T? {
     semaphoreQueue.wait()
     let element: T? = list.last
     semaphoreQueue.signal()
     return element
   }
 
-  public func clear() {
+  func clear() {
     semaphoreQueue.wait()
     list.removeAll()
     semaphoreQueue.signal()
   }
 }
 
-public class FailureQueue: Collection<UInt64> {
-
-  public var currentTimeWindow: UInt64? {
+class FailureQueue: Collection<UInt64> {
+  var currentTimeWindow: UInt64? {
     semaphoreQueue.wait()
     // Get time difference
     let timeWindow: UInt64?
@@ -106,5 +105,62 @@ extension Date {
   public static func currentTimeMillis() -> UInt64 {
     let timeInMillis = UInt64(NSDate().timeIntervalSince1970 * 1000.0)
     return timeInMillis
+  }
+}
+
+class Bulkhead {
+  private let serialQueue: DispatchQueue
+  private let concurrentQueue: DispatchQueue
+  private let semaphore: DispatchSemaphore
+
+  init(limit: Int) {
+    serialQueue = DispatchQueue(label: "bulkheadSerialQueue")
+    concurrentQueue = DispatchQueue(label: "bulkheadConcurrentQueue", attributes: .concurrent)
+    semaphore = DispatchSemaphore(value: limit)
+  }
+
+  func enqueue(task: @escaping () -> Void ) {
+    serialQueue.async { [weak self] in
+      self?.semaphore.wait()
+      self?.concurrentQueue.async {
+        task()
+        self?.semaphore.signal()
+      }
+    }
+  }
+}
+
+// Invocation entity
+public class Invocation<A, B, C> {
+
+  public let commandArgs: A
+  private(set) var timedOut: Bool = false
+  private(set) var completed: Bool = false
+  weak private var breaker: CircuitBreaker<A, B, C>?
+  public init(breaker: CircuitBreaker<A, B, C>, commandArgs: A) {
+    self.commandArgs = commandArgs
+    self.breaker = breaker
+  }
+
+  public func setTimedOut() {
+    self.timedOut = true
+  }
+
+  public func setCompleted() {
+    self.completed = true
+  }
+
+  public func notifySuccess() {
+    if !self.timedOut {
+      self.setCompleted()
+      breaker?.notifySuccess()
+    }
+  }
+
+  public func notifyFailure() {
+    if !self.timedOut {
+      self.setCompleted()
+      breaker?.notifyFailure()
+    }
   }
 }
